Everything is working now
Have been working on this one issue with the catalogue not showing the updated text via the adminEnterBookUpdate class. The signals and slots were correct but it seemed that it was creating a new instance as I navigated through the application. After many hours this has been resolved! I had to initialise the two classes that interact with the signals and slots in a separate GlobalInstances class file, ensuring the correct instance is called when opening the pages throughout the navigation of the application.